requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
Flask-Migrate
alembic
openpyxl
reportlab
weasyprint
